+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
-# intellij-spring-green-light-theme Changelog
+# intellij-mint-green-light-theme Changelog
+
+## 1.0.0 - 2024-12-17
+- rename "spring" to "mint"
 
 ## 1.0.0 - 2024-02-20
 - initial release